+++ app/src/main/java/woowacourse/payments/ui/cardwallet/components/CardWalletCards.kt
@@ -7,8 +7,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import woowacourse.payments.designsystem.theme.AndroidpaymentsTheme
+import woowacourse.payments.ui.common.components.PaymentCard
 import woowacourse.payments.ui.common.model.CardUiModel
-import woowacourse.payments.ui.common.model.components.PaymentCard
 
 @Composable
 fun CardWalletCards(cards: List<CardUiModel>) {
